% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{detect_pattern}
\alias{detect_pattern}
\title{Detects a population pattern in given result from a sum calibration.}
\usage{
detect_pattern(sum_cal_result, conf_envelope = NULL)
}
\arguments{
\item{sum_cal_result}{the sum calibration result}

\item{conf_envelope}{optional a confindence envelope}
}
\value{
true or false if a pattern was detected or not
}
\description{
Detects a population pattern in given result from a sum calibration.
}
