% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation_functions.R
\name{prepare_data}
\alias{prepare_data}
\title{Turns the input data into a probability distribution and hands this and
other parameters to further analysis}
\usage{
prepare_data(this_data)
}
\arguments{
\item{this_data}{the raw population estimate data}
}
\value{
a list containing the number of dates and the signal strengths
from the input data, and a probability distribution derived from the population data
}
\description{
Turns the input data into a probability distribution and hands this and
other parameters to further analysis
}
