% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{read_input_data}
\alias{read_input_data}
\title{Read the three input files and turns them into a object (list)
for further analysis}
\usage{
read_input_data(path = "analysis/data/raw_data", test = FALSE)
}
\arguments{
\item{path}{The location of the input files}

\item{test}{boolean if this is a test run or not}
}
\value{
a list containing the data frames from the imported csv files
}
\description{
Read the three input files and turns them into a object (list)
for further analysis
}
