% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_functions.R
\name{set_up_environment}
\alias{set_up_environment}
\title{Prepares the environment for further analysis by loading
and initialising necessary packages}
\usage{
set_up_environment()
}
\description{
Prepares the environment for further analysis by loading
and initialising necessary packages
}
