% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_dates}
\alias{simulate_dates}
\title{Samples a number of calendar dates from a given population (probability)
distribution according to a given signal strength.}
\usage{
simulate_dates(this_data, n_date, signal_strength)
}
\arguments{
\item{this_data}{The input data set}

\item{n_date}{The number of calendar dates to be sampled}

\item{signal_strength}{The signal strength of the pattern}
}
\value{
A list of calendar dates with a density according
to the population (probability) distribution.
}
\description{
Samples a number of calendar dates from a given population (probability)
distribution according to a given signal strength.
}
