% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{smooth_sumcal_result}
\alias{smooth_sumcal_result}
\title{Smooths the result of a sum calibration with a moving average in a given window}
\usage{
smooth_sumcal_result(x, window_size)
}
\arguments{
\item{x}{The result of a sum calibration}

\item{window_size}{The size of the window of the moving average}
}
\value{
a smoothed sum calibration result
}
\description{
Smooths the result of a sum calibration with a moving average in a given window
}
