% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{sum_calibrate_simulated_dates}
\alias{sum_calibrate_simulated_dates}
\title{Creates a sum calibration from a number of dates with OxCals C_Simulate function}
\usage{
sum_calibrate_simulated_dates(this_dates)
}
\arguments{
\item{this_dates}{A number of calendar dates}
}
\value{
The result of the sum calibration
}
\description{
Creates a sum calibration from a number of dates with OxCals C_Simulate function
}
